refactor: changes for build
