+++ apps/commudle-admin/src/app/feature-modules/community-channels/components/community-forum-list/community-forum-list.component.ts
@@ -117,6 +117,7 @@ export class CommunityForumListComponent implements OnInit, OnDestroy {
     this.dialogService.open(NewCommunityChannelComponent, {
       closeOnBackdropClick: false,
       hasBackdrop: false,
+      hasScroll: true,
       context: {
         groupName: groupName,
         discussionType: this.discussionType.FORUM,
