+++ apps/commudle-admin/src/app/feature-modules/community-channels/services/community-channels.service.ts
@@ -1,225 +1,3 @@
-/* eslint-disable @nrwl/nx/enforce-module-boundaries */
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
-import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
-import { HttpClient, HttpParams } from '@angular/common/http';
-import { ApiRoutesService } from 'apps/shared-services/api-routes.service';
-import { ICommunityChannel } from 'apps/shared-models/community-channel.model';
-import { API_ROUTES } from 'apps/shared-services/api-routes.constants';
-import { ICommunityChannels } from 'apps/shared-models/community-channels.model';
-import { IUserRolesUsers } from 'apps/shared-models/user_roles_users.model';
-import { IUsers } from 'apps/shared-models/users.model';
-import { ICommunities } from 'apps/shared-models/communities.model';
-import { IUserMessages } from 'apps/shared-models/user_messages.model';
-import { IUserMessage } from 'apps/shared-models/user_message.model';
-import { IPagination } from 'apps/shared-models/pagination.model';
-import { IUserRolesUser } from '@commudle/shared-models';
+import { CommunityChannelsService } from '@commudle/shared-services';
 
-@Injectable({
-  providedIn: 'root',
-})
-export class CommunityChannelsService {
-  constructor(private http: HttpClient, private apiRoutesService: ApiRoutesService) {}
-
-  getChannelInfo(id): Observable<ICommunityChannel> {
-    const params = new HttpParams().set('community_channel_id', id);
-    return this.http.get<any>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.SHOW), { params });
-  }
-
-  getUserChannelCommunities(): Observable<ICommunities> {
-    return this.http.get<any>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.USER_CHANNEL_COMMUNITIES));
-  }
-
-  create(communityId, channelData): Observable<ICommunityChannel> {
-    const params = new HttpParams().set('community_id', communityId);
-    return this.http.post<ICommunityChannel>(
-      this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.CREATE),
-      channelData,
-      { params },
-    );
-  }
-
-  update(communityChannelId, channelData): Observable<ICommunityChannel> {
-    const params = new HttpParams().set('community_channel_id', communityChannelId);
-    return this.http.put<ICommunityChannel>(
-      this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.UPDATE),
-      channelData,
-      { params },
-    );
-  }
-
-  delete(communityChannelId, archive): Observable<boolean> {
-    const params = new HttpParams().set('community_channel_id', communityChannelId).set('archive', archive);
-
-    return this.http.delete<boolean>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.DELETE), { params });
-  }
-
-  index(communityId, displayType): Observable<IPagination<ICommunityChannels>> {
-    const params = new HttpParams().set('community_id', communityId).set('display_type', displayType).set('limit', 50);
-    return this.http.get<IPagination<ICommunityChannels>>(
-      this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.INDEX),
-      {
-        params,
-      },
-    );
-  }
-
-  getJoinToken(communityChannelId): Observable<string> {
-    const params = new HttpParams().set('community_channel_id', communityChannelId);
-    return this.http.get<string>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.JOIN_TOKEN), { params });
-  }
-
-  resetJointoken(communityChannelId): Observable<string> {
-    return this.http.put<string>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.RESET_JOIN_TOKEN), {
-      community_channel_id: communityChannelId,
-    });
-  }
-
-  inviteMembers(communityChannelId, userRoleData): Observable<IUserRolesUser> {
-    const params = new HttpParams().set('community_channel_id', communityChannelId);
-    return this.http.post<IUserRolesUser>(
-      this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.MEMBERS.INVITE),
-      {
-        user_roles_user: userRoleData,
-      },
-      { params },
-    );
-  }
-
-  joinChannel(communityChannelId: number, token?: string, decline?: boolean): Observable<boolean> {
-    const params = {} as any;
-    params.community_channel_id = communityChannelId;
-    if (token) {
-      params.token = token;
-    }
-    if (decline) {
-      params.decline = decline;
-    }
-    return this.http.put<boolean>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.JOIN_CHANNEL), params);
-  }
-
-  joinByToken(token, decline): Observable<any> {
-    return this.http.post<any>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.MEMBERS.JOIN_BY_TOKEN), {
-      token,
-      decline: decline,
-    });
-  }
-
-  membersList(communityChannelId, page, count): Observable<IUserRolesUsers> {
-    const params = new HttpParams()
-      .set('community_channel_id', communityChannelId)
-      .set('page', page)
-      .set('count', count);
-    return this.http.get<IUserRolesUsers>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.MEMBERS.INDEX), {
-      params,
-    });
-  }
-
-  toggleAdmin(userRolesUserId): Observable<IUserRolesUser> {
-    return this.http.put<IUserRolesUser>(
-      this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.MEMBERS.TOGGLE_ADMIN),
-      {
-        user_roles_user_id: userRolesUserId,
-      },
-    );
-  }
-
-  removeMembership(userRolesUserId): Observable<any> {
-    const params = new HttpParams().set('user_roles_user_id', userRolesUserId);
-    return this.http.delete<any>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.MEMBERS.REMOVE), {
-      params,
-    });
-  }
-
-  exitChannel(channelId): Observable<any> {
-    const params = new HttpParams().set('community_channel_id', channelId);
-    return this.http.delete<any>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.MEMBERS.EXIT_CHANNEL), {
-      params,
-    });
-  }
-
-  getTaggableUsers(query, channelId): Observable<IUsers> {
-    const params = new HttpParams().set('query', query).set('community_channel_id', channelId);
-    return this.http.get<IUsers>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.TAGGABLE_USERS), {
-      params,
-    });
-  }
-
-  deleteLogo(communityChannelId): Observable<boolean> {
-    const params = new HttpParams().set('community_channel_id', communityChannelId);
-
-    return this.http.delete<boolean>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.DELETE_LOGO), {
-      params,
-    });
-  }
-
-  getDiscussionMessages(communityChannelId, page, count): Observable<IUserMessages> {
-    const params = new HttpParams()
-      .set('community_channel_id', communityChannelId)
-      .set('page', page)
-      .set('count', count);
-    return this.http.get<IUserMessages>(
-      this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.DISCUSSION_MESSAGES),
-      { params },
-    );
-  }
-
-  getDiscussionMessagesForScroll(communityChannelId, count, action, messageId): Observable<IUserMessages> {
-    let params;
-    if (action === 'initial') {
-      params = new HttpParams().set('community_channel_id', communityChannelId).set('count', count);
-    } else {
-      params = new HttpParams()
-        .set('community_channel_id', communityChannelId)
-        .set('type', action)
-        .set('user_message_id', messageId)
-        .set('count', count);
-    }
-    return this.http.get<IUserMessages>(
-      this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.DISCUSSION_MESSAGES_SCROLL),
-      { params },
-    );
-  }
-
-  sendMessageByEmail(userMessageId, communityChannelId): Observable<boolean> {
-    return this.http.post<boolean>(
-      this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.SEND_MESSAGE_BY_EMAIL_TO_ALL_MEMBERS),
-      {
-        community_channel_id: communityChannelId,
-        user_message_id: userMessageId,
-      },
-    );
-  }
-
-  pinMessage(userMessageId, communityChannelId): Observable<boolean> {
-    return this.http.post<boolean>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.PINNING_MESSAGES.PIN), {
-      community_channel_id: communityChannelId,
-      user_message_id: userMessageId,
-    });
-  }
-
-  unpinMessage(userMessageId, communityChannelId): Observable<boolean> {
-    return this.http.post<boolean>(
-      this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.PINNING_MESSAGES.UNPIN),
-      {
-        community_channel_id: communityChannelId,
-        user_message_id: userMessageId,
-      },
-    );
-  }
-
-  getPinnedMessages(communityChannelId): Observable<IUserMessage[]> {
-    const params = new HttpParams().set('community_channel_id', communityChannelId);
-    return this.http.get<IUserMessage[]>(
-      this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.PINNING_MESSAGES.PINNED_MESSAGES),
-      { params },
-    );
-  }
-
-  showByToken(token): Observable<any> {
-    const params = new HttpParams().set('token', token);
-    return this.http.get<any>(this.apiRoutesService.getRoute(API_ROUTES.COMMUNITY_CHANNELS.SHOW_BY_TOKEN), {
-      params,
-    });
-  }
-}
+export { CommunityChannelsService };
