+++ apps/shared-models/community-channel.model.ts
@@ -1,24 +1,3 @@
-import { IAttachedFile } from './attached-file.model';
-import { IUser } from './user.model';
+import { ICommunityChannel } from '@commudle/shared-models';
 
-export interface ICommunityChannel {
-  id: number;
-  user: IUser;
-  kommunity_id: number;
-  name: string;
-  description: string;
-  group_name: string;
-  join_token: string;
-  is_private: boolean;
-  is_readonly: boolean;
-  logo: IAttachedFile;
-  my_roles: any[];
-  member_count: number;
-  messages_count?: number;
-  display_type: string;
-  messages_count_in_three_months?: number;
-  kommunity?: {
-    id: number;
-    name: string;
-  };
-}
+export { ICommunityChannel };
