+++ libs/shared/components/src/lib/components/channel-discussion/channel-discussion.component.html
@@ -75,7 +75,7 @@
     </a>
   </ng-template>
   <ng-template #joinButton>
-    <a (click)="joinChannelButton()" nbButton status="primary" fullWidth>
+    <a (click)="joinChannel()" nbButton status="primary" fullWidth>
       <nb-icon icon="log-in-outline"></nb-icon>
       Join This Channel
     </a>
