+++ libs/shared/models/src/index.ts
@@ -12,3 +12,5 @@ export * from './lib/user.model';
 export * from './lib/user-message.model';
 export * from './lib/user-role.model';
 export * from './lib/user-roles-user.model';
+export * from './lib/community-channel.model';
+export * from './lib/user_roles.enum';
